requirements.txt
langchain
ffmpeg-python
pydantic
langchain-community